pyproject.toml
[tool.poetry]
name = "capstone-project"
version = "0.1.0"
description = ""
authors = ["lambdacore <jonathanwh93@gmail.com>"]
readme = "README.md"
packages = [{include = "capstone_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
numpy = "^1.23.4"
django-tailwind = "^3.4.0"
django-browser-reload = "^1.6.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
pymysql = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
